requirements.txt
# === Core Logic & AI Agents ===
openai==2.3.0
anthropic>=0.39.0
trueskill==0.4.5
json_repair==0.53.0
tenacity==9.1.2

# === Web Crawling & APIs ===
requests==2.32.5
beautifulsoup4==4.14.2
google-search-results>=2.4.2
httpx==0.28.1
urllib3==2.5.0

# === Data Science & Math ===
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2

# === Visualization ===
matplotlib==3.10.7
seaborn==0.13.2
adjustText==1.3.0
pillow==12.0.0
# networkx>=3.1

# === Utilities ===
filelock>=3.12.0
python-dotenv>=1.0.0
tqdm==4.67.1
python-dateutil==2.9.0.post0
pytz==2025.2
openpyxl==3.1.5  # Excel support for pandas

# === Dependencies for reproducibility ===
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fonttools==4.61.0
h11==0.16.0
httpcore==1.0.9
idna==3.11
jiter==0.11.0
joblib==1.5.2
kiwisolver==1.4.9
packaging==25.0
pydantic==2.12.2
pydantic_core==2.41.4
pyparsing==3.2.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2